Handle non-string lists in query arguments
Allow any type that can be coerced into a string to be used as a vector
of arguments for a query. For example, the API endpoint
'questions/{ids}' takes a semicolon-delimited vector of ids.

This can be passed as a string:

    stackexchange.query(
        'vi',
        stackexchange.queries.questions.by_id.ALL,
        ids = "42;128"
    )

Or, as a list of strings:

    stackexchange.query(
        'vi',
        stackexchange.queries.questions.by_id.ALL,
        ids = ["42", "128"]
    )

Or, as a list of type T, where T is a type that can be coerced to a
string with str():

    stackexchange.query(
        'vi',
        stackexchange.queries.questions.by_id.ALL,
        ids = [42, 128]
    )

These are all semantically identical, and will result in the same
formatted endpoint of 'questions/42;128?site=vi'