Consolidate structures
Make `DictTree` a more abstract `LookupDict` type, as it more accurately
represents the purpose of the class.

A `LookupDict` is similar to an ordinary `dict` object. The key behavior
change is that attempting to access an undefined attribute of a
`LookupDict` returns `None` instead of raising an `AttributeError`,
effectively making `dict.get(key, None)` the default key access
behavior.