Make URLTree __init__ take a list of method dicts
URL method lookup is now self contained in the URLTree class. This is
done by passing a list of tuples, each containing a HTTP method string
and a URL dictionary. For example:

    url_dict = {
        "GET": {
            # URLTrees can be nested
            "questions": {
                "all": "questions",
                "featured": "questions/featured"
            "comments": {
                "all": "comments"
            }
        },
        "POST": {
            "auth": "authorize",
            "questions": {
                "add": "questions/add"
            },
            "comments": {
                "add": "comments/add",
                "edit:" "comments/{id}/edit"
            }
        }
    }

    queries = URLTree([method, urls for k, v in url_dict])

This allows the constructor to initialize an internal `_method` lookup
table. This can be accessed directly, or by calling the `method` method
of the URLTree class:

    print(queries._method.comments.ADD) # POST
    print(queries.method(queries.questions.FEATURED) # GET

This has the same effect as the previous implementation (synthesizing
larger dictionaries by merging smaller ones), but allows for a nice
conciseness in combination with list comprehension (shown above).